+++ plugins/semanticcache/changelog.md
@@ -1,4 +1,5 @@
 <!-- The pattern we follow here is to keep the changelog for the latest version -->
 <!-- Old changelogs are automatically attached to the GitHub releases -->
 
-- chore: version update core to 1.2.12 and framework to 1.1.14
+- chore: version update core to 1.2.13 and framework to 1.1.15
+- tests: added mocker plugin to all chat/responses tests